Refactored code and updated requirements
