+++ data/TPI_sector_data_All_sectors_08032025/User guide TPI files.xlsx
