+++ requirements.txt
@@ -1,80 +1,38 @@
 annotated-types==0.7.0
-anyio==4.8.0
-appnope==0.1.4
-asttokens==3.0.0
-bcrypt==3.2.0
-certifi==2025.1.31
-cffi==1.17.1
-charset-normalizer==3.4.1
+anyio==4.9.0
+bcrypt==4.3.0
 click==8.1.8
-colorlog==6.8.2
-comm==0.2.2
-cryptography==45.0.2
-debugpy==1.8.14
-decorator==5.2.1
 Deprecated==1.2.18
-dot==0.3.0
+dotenv==0.9.9
 ecdsa==0.19.1
-env==0.1.0
 et_xmlfile==2.0.0
-executing==2.2.0
-fastapi==0.115.11
-h11==0.14.0
-httpcore==1.0.7
-httpx==0.28.1
+exceptiongroup==1.3.0
+fastapi==0.115.12
+greenlet==3.2.2
+h11==0.16.0
 idna==3.10
-iniconfig==2.0.0
-ipykernel==6.29.5
-ipython==9.2.0
-ipython_pygments_lexers==1.1.1
-jedi==0.19.2
-jupyter_client==8.6.3
-jupyter_core==5.7.2
-limits==5.2.0
-matplotlib-inline==0.1.7
-nest-asyncio==1.6.0
-numpy==2.2.3
+limits==4.2
+numpy==2.0.2
 openpyxl==3.1.5
 packaging==24.2
 pandas==2.2.3
-parso==0.8.4
-pathlib==1.0.1
-pexpect==4.9.0
-platformdirs==4.3.8
-pluggy==1.5.0
-prompt_toolkit==3.0.51
-psutil==7.0.0
-ptyprocess==0.7.0
-pure_eval==0.2.3
+psycopg2==2.9.10
 pyasn1==0.4.8
-pycountry==24.6.1
-pycparser==2.22
-pydantic==2.10.6
-pydantic_core==2.27.2
-Pygments==2.19.1
-pytest==8.3.5
-pytest-snapshot==0.9.0
+pydantic==2.11.5
+pydantic_core==2.33.2
 python-dateutil==2.9.0.post0
-python-dotenv==1.0.0
+python-dotenv==1.1.0
 python-jose==3.4.0
-python-multipart==0.0.6
-pytz==2025.1
-pyzmq==26.4.0
-requests==2.32.3
+python-multipart==0.0.20
+pytz==2025.2
 rsa==4.9.1
 six==1.17.0
 slowapi==0.1.9
 sniffio==1.3.1
-sqlalchemy==2.0.30
-stack-data==0.6.3
-starlette==0.46.1
-tornado==6.5.1
-tqdm==4.67.1
-traitlets==5.14.3
-typing_extensions==4.12.2
-tzdata==2025.1
-urllib3==2.3.0
-uvicorn==0.34.0
-wcwidth==0.2.13
+SQLAlchemy==2.0.41
+starlette==0.46.2
+typing-inspection==0.4.1
+typing_extensions==4.13.2
+tzdata==2025.2
+uvicorn==0.34.2
 wrapt==1.17.2
-psycopg2-binary==2.9.9
